%% the data is extract from Package golubEsets.

\name{Golub_Merge}
\alias{Golub_Merge}

\title{Gene expression data from Alon et al. (1999)}
\usage{
data(Golub_Merge)
}

\description{
Golub_Merge data  
}


\value{
  \item{X}{X}
  \item{y}{y} 
}



\references{
Alon, U. and Barkai, N. and Notterman, D.A. and Gish, K. and Ybarra, S. and Mack, D. and Levine, A.J. (1999).
Broad patterns of gene expression revealed by clustering analysis of tumor and normal colon tissues probed by oligonucleotide arrays,
Proc. Natl. Acad. Sci. USA,\bold{96}(12), 6745--6750.
}


\examples{
# load plsgenomics library
library(SubLasso)

# load data set
data(Golub_Merge)

}


\keyword{datasets}
