\name{SubLasso-package}
\alias{SubLasso-package}
\docType{package}
\title{
SubLasso package
}
\description{
SubLasso package
}
\details{
\tabular{ll}{
Package: \tab SubLasso\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-10-01\cr
License: \tab GPL-2 \cr
}
A convient procedure for microarray studies. It can do feature selection and classification simultaneously for binary outcomes . K-folds cross validation results were returned for users. Users needn't to adjust the tune parameter and can fix any features that they desire to keep in the model.
}
\author{
Author: Fengfeng Zhou, Youxi Luo, Qinghan Meng, Ruiquan Ge, Guoqin Mai, Jikui Liu\cr
Maintainer: Fengfeng Zhou <ff.zhou@siat.ac.cn>\cr
}
\references{
1.	Smyth GK: Linear models and empirical bayes methods for assessing differential expression in microarray experiments. Statistical applications in genetics and molecular biology 2005, 3:Article3.\cr
2.	Tusher VG, Tibshirani R, Chu G: Significance analysis of microarrays applied to the ionizing radiation response. Proceedings of the National Academy of Sciences of the United States of America 2001, 98(9):5116-5121.\cr
3.	Culhane AC, Perriere G, Considine EC, Cotter TG, Higgins DG: Between-group analysis of microarray data. Bioinformatics 2002, 18(12):1600-1608.\cr
4.	Parodi S, Muselli M, Fontana V, Bonassi S: ROC curves are a suitable and flexible tool for the analysis of gene expression profiles. Cytogenetic and genome research 2003, 101(1):90-91.\cr
5.	Wang Q, Wei F, Lv G, Li C, Liu T, Hadjipanayis CG, Zhang G, Hao C, Bellail AC: The association of TP53 mutations with the resistance of colorectal carcinoma to the insulin-like growth factor-1 receptor inhibitor picropodophyllin. Bmc Cancer 2013, 13(1):521.\cr
6.	Cortot AB, Younes M, Martel-Planche G, Guibert B, Isaac S, Souquet PJ, Commo F, Girard P, Fouret P, Brambilla E et al: Mutation of TP53 and Alteration of p14 Expression in EGFR- and KRAS-Mutated Lung Adenocarcinomas. Clinical lung cancer 2013.\cr
7.	Singamsetty GK, Malempati S, Bhogadhi S, Kondreddy R, Govatati S, Tangudu NK, Govatati S, Kuraganti AK, Bhanoori M, Kassetty K: TP53 alterations and colorectal cancer predisposition in south Indian population: A case-control study. Tumour biology : the journal of the International Society for Oncodevelopmental Biology and Medicine 2013.\cr
8.	Chung R, Whaley J, Kley N, Anderson K, Louis D, Menon A, Hettlich C, Freiman R, Hedley-Whyte ET, Martuza R et al: TP53 gene mutations and 17p deletions in human astrocytomas. Genes, chromosomes & cancer 1991, 3(5):323-331.\cr
9.	Eeles RA, Bartkova J, Lane DP, Bartek J: The role of TP53 in breast cancer development. Cancer surveys 1993, 18:57-75.\cr
10.	Ding J, Shi J, Wu FX: SVM-RFE based feature selection for tandem mass spectrum quality assessment. International journal of data mining and bioinformatics 2011, 5(1):73-88.\cr
11.	Mundra PA, Rajapakse JC: SVM-RFE with MRMR filter for gene selection. IEEE transactions on nanobioscience 2010, 9(1):31-37.\cr
12.	Tang Y, Zhang YQ, Huang Z: Development of two-stage SVM-RFE gene selection strategy for microarray expression data analysis. IEEE/ACM transactions on computational biology and bioinformatics / IEEE, ACM 2007, 4(3):365-381.\cr
13.	Zhou X, Tuck DP: MSVM-RFE: extensions of SVM-RFE for multiclass gene selection on DNA microarray data. Bioinformatics 2007, 23(9):1106-1114.\cr
14.	Niijima S, Kuhara S: Recursive gene selection based on maximum margin criterion: a comparison with SVM-RFE. BMC bioinformatics 2006, 7:543.\cr
15.	Ding Y, Wilkins D: Improving the performance of SVM-RFE to select genes in microarray data. BMC bioinformatics 2006, 7 Suppl 2:S12.\cr
16.	Duan KB, Rajapakse JC, Wang H, Azuaje F: Multiple SVM-RFE for gene selection in cancer classification with expression data. IEEE transactions on nanobioscience 2005, 4(3):228-234.\cr
17.	Tibshirani R: Regression shrinkage and selection via the Lasso. J Roy Stat Soc B Met 1996, 58(1):267-288.\cr
18.	Friedman J, Hastie T, Tibshirani R: Regularization Paths for Generalized Linear Models via Coordinate Descent. J Stat Softw 2010, 33(1):1-22.\cr
}
\keyword{ package }
